Cardano.Pool.Jormungandr.Metrics

Description

This module can fold over a blockchain to collect metrics about Stake pools.

It interacts with: - Cardano.Wallet.Network which provides the chain - Cardano.Pool.DB - which can persist the metrics - Cardano.Wallet.Api.Server - which presents the results in an endpoint

Types

data Block

Information from a block relevant to monitoring stake pools.

Constructors

Block

Fields

header :: BlockHeader

producer :: PoolId

The stake pool that minted this block.

poolRegistrations :: ![PoolRegistrationCertificate]

Any stake pools that were registered in this block.

Listing stake-pools from the DB

newStakePoolLayer

Arguments

:: Tracer IO StakePoolLog

-> BlockHeader
Genesis block header

-> (EpochNo -> Quantity "lovelace" Word64 -> EpochConstants)

-> DBLayer IO

-> NetworkLayer IO t Block

-> FilePath
A directory to cache downloaded stake pool metadata. Will be created if it does not exist.

-> StakePoolLayer ErrListStakePools IO

data ErrListStakePools

Constructors

ErrMetricsIsUnsynced (Quantity "percent" Percentage)

ErrListStakePoolsCurrentNodeTip ErrCurrentNodeTip

Following the chain

monitorStakePools

Arguments

:: Tracer IO StakePoolLog

-> (Block, Quantity "block" Word32)
Genesis block and k

-> NetworkLayer IO t Block

-> DBLayer IO

-> IO ()

monitorStakePools follows the chain and puts pool productions and stake distributions to a DBLayer, such that the data in the database is always consistent.

The pool productions and stake distrubtions in the db can never be from different forks such that it's safe for readers to access it.

Combining Metrics

combineMetrics :: Map PoolId (Quantity "lovelace" Word64) -> Map PoolId (Quantity "block" Word64) -> Map PoolId Double -> Map PoolId (Quantity "lovelace" Word64, Quantity "block" Word64, Double)

Combines three different sources of data into one:

A stake-distribution map

A pool-production map

A pool-performance map

If a pool is in 2 or 3 but not in 1, it means that the pool has been de-registered.

If a pool is in (1+2) but not in (3), it simply means it has produced 0 blocks so far.

Associating metadata

associateMetadata

Arguments

:: [(PoolId, [PoolOwner])]
Ordered mapping from pool to owner(s).

-> [(PoolOwner, Maybe StakePoolMetadata)]
Association between owner and metadata

-> [(StakePoolLog, Maybe StakePoolMetadata)]

Given a mapping from PoolId -> PoolOwner and a mapping between PoolOwner - StakePoolMetadata, return a matching StakePoolMeta entry for every PoolId.

If there is no metadata for a pool, it returns Nothing for that PoolId. If there is different metadata submitted by multiple owners of a pool, it returns Nothing. If there is one unique metadata for a pool, it returns Just the metadata for that PoolId.

It also provides a log message for each association.

Logging

data StakePoolLog

Messages associated with stake pool layer.

Constructors

MsgRegistry RegistryLog

MsgListStakePoolsBegin

MsgMetadataUnavailable

MsgMetadataUsing PoolId PoolOwner StakePoolMetadata

MsgMetadataMissing PoolId

MsgMetadataMultiple PoolId [(PoolOwner, StakePoolMetadata)]

MsgComputedProgress BlockHeader BlockHeader

MsgStartMonitoring [BlockHeader]

MsgFollow FollowLog

MsgStakeDistribution EpochNo

MsgStakePoolRegistration PoolRegistrationCertificate

MsgRollingBackTo SlotId

MsgApplyError ErrMonitorStakePools

MsgUsingRankingEpochConstants EpochConstants

MsgUsingTotalStakeForRanking (Quantity "lovelace" Word64)