Cardano.Wallet.Binary.Jormungandr

Description

The format is for the Shelley era as implemented by the Jörmungandr node.

Documentation

getBlockHeader :: Get BlockHeader

getBlock :: Get Block

data Message

Messages are what the block body consists of.

Every message is prefixed with a message header.

Following, as closely as possible: https://github.com/input-output-hk/rust-cardano/blob/e0616f13bebd6b908320bddb1c1502dea0d3305a/chain-impl-mockchain/src/message/mod.rs#L22-L29

Constructors

Initial [ConfigParam]

OldUtxoDeclaration TODO

Transaction TODO

Certificate TODO

UpdateProposal SignedUpdateProposal

UpdateVote SignedVote

UnimplementedMessage Int

data BlockHeader

Constructors

BlockHeader

Fields

version :: Word16

contentSize :: Word32

slot :: SlotId

chainLength :: Word32

contentHash :: Hash "content"

parentHeaderHash :: Hash "parentHeader"

Re-export

runGet :: Get a -> ByteString -> a

The simplest interface to run a Get decoder. If the decoder runs into an error, calls fail, or runs out of input, it will call error.