Cardano.Wallet.Compatibility.Jormungandr

Description

Contains various implementation decision that are specific to a particular network / protocol. This allows us to easily select a particular backend (Byron, Shelley-Rust, Shelley-Haskell) and isolate the bits that vary between those backends.

Target

data Jormungandr

A type representing the Jormungandr as a network target. This has an influence on binary serializer & network primitives. See also TxId