Cardano.Wallet.Jormungandr.Api

Description

An specification for the Jörmungandr REST API.

Documentation

type Api = GetTipId :<|> (GetBlock :<|> (GetBlockDescendantIds :<|> (PostMessage :<|> GetStakeDistribution)))

newtype ApiT a

Polymorphic wrapper type to put around primitive types and, 3rd party lib types to avoid defining orphan instances and/or, undesirable instances on primitive types. It helps to keep a nice separation of concerns between the API layer and other modules.

Constructors

ApiT

Fields

getApiT :: a

type GetBlock = "api" :> ("v0" :> ("block" :> (Capture "blockHeaderHash" BlockId :> Get '[JormungandrBinary] Block)))

Retrieve a block by its id.

type GetTipId = "api" :> ("v0" :> ("tip" :> Get '[Hex] BlockId))

Retrieve the header of the latest known block.

type GetBlockDescendantIds = "api" :> ("v0" :> ("block" :> (Capture "blockId" BlockId :> ("next_id" :> (QueryParam "count" Word :> Get '[JormungandrBinary] [BlockId])))))

Retrieve n descendants of a given block, sorted from closest to farthest.

There might also exist fewer than n descendants.

For n=3 we might have:

[genesis] ... -- [b] -- [b+1] -- [b+2] -- [b+3] -- ... -- [tip]
                  \       \                  \
                 parent    +--- descendants ---+

type GetStakeDistribution = "api" :> ("v0" :> ("stake" :> Get '[JSON] StakeApiResponse))

Retrieve stake distribution

type PostMessage = "api" :> ("v0" :> ("message" :> (ReqBody '[JormungandrBinary] (Tx, [TxWitness]) :> Post '[NoContent] NoContent)))

data StakeApiResponse

Constructors

StakeApiResponse

Fields

epoch :: ApiT EpochNo

stake :: ApiStakeDistribution

data ApiStakeDistribution

Constructors

ApiStakeDistribution

Fields

dangling :: ApiT (Quantity "lovelace" Word64)

pools :: [(ApiT PoolId, ApiT (Quantity "lovelace" Word64))]

unassigned :: ApiT (Quantity "lovelace" Word64)

newtype BlockId

Constructors

BlockId

Fields

getBlockId :: Hash "BlockHeader"

api :: Proxy Api