Cardano.Wallet.Jormungandr.Api.Client

Description

This module implements parts of a Jörmungandr REST API Client. It also provides some top-level helpers that are built using the API and are quite handy in other situations.

Documentation

data JormungandrClient m

Endpoints of the jormungandr REST API.

Constructors

JormungandrClient

Fields

getTipId :: ExceptT ErrNetworkUnavailable m (Hash "BlockHeader")

getBlock :: Hash "BlockHeader" -> ExceptT ErrGetBlock m Block

getDescendantIds :: Hash "BlockHeader" -> Word -> ExceptT ErrGetDescendants m [Hash "BlockHeader"]

postMessage :: (Tx, [TxWitness]) -> ExceptT ErrPostTx m ()

getInitialBlockchainParameters :: Hash "Genesis" -> ExceptT ErrGetBlockchainParams m (Block, BlockchainParameters)

getStakeDistribution :: ExceptT ErrNetworkUnavailable m StakeApiResponse

mkJormungandrClient :: Manager -> BaseUrl -> JormungandrClient IO

Construct a JormungandrClient

Extra Helpers

getBlockHeader :: Monad m => JormungandrClient m -> Hash "BlockHeader" -> ExceptT ErrGetBlock m BlockHeader

Get a block header corresponding to a header hash.

getBlocks

Arguments

:: Monad m

=> JormungandrClient m

-> Word
Batch size to be used for fetching blocks

-> Hash "BlockHeader"
Block ID to start from

-> ExceptT ErrGetBlock m [Block]

Errors

class LiftError lift where

Safely lift error into a bigger type.

Methods

liftE :: lift

data ErrGetBlock

Error while trying to get one or more blocks

Constructors

ErrGetBlockNetworkUnreachable ErrNetworkUnavailable

ErrGetBlockNotFound (Hash "BlockHeader")

data ErrGetBlockchainParams

Constructors

ErrGetBlockchainParamsNetworkUnreachable ErrNetworkUnavailable

ErrGetBlockchainParamsGenesisNotFound (Hash "Genesis")

ErrGetBlockchainParamsIncompleteParams [ConfigParam]

data ErrGetDescendants

Constructors

ErrGetDescendantsNetworkUnreachable ErrNetworkUnavailable

ErrGetDescendantsParentNotFound (Hash "BlockHeader")

data ErrNetworkTip

Error while trying to get the network tip

Constructors

ErrNetworkTipNetworkUnreachable ErrNetworkUnavailable

ErrNetworkTipNotFound

data ErrNetworkUnavailable

Network is unavailable

Constructors

ErrNetworkUnreachable Text
Cannot connect to network backend.

ErrNetworkInvalid Text
Network backend reports that the requested network is invalid.

data ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxNetworkUnreachable ErrNetworkUnavailable

ErrPostTxBadRequest Text

ErrPostTxProtocolFailure Text

data ErrUnexpectedNetworkFailure

Constructors

ErrUnexpectedNetworkFailure Link ServantError

Re-export

data BaseUrl

Constructors

BaseUrl

Fields

baseUrlScheme :: Scheme

baseUrlHost :: String

baseUrlPort :: Int

baseUrlPath :: String

data Scheme

Constructors

Http

Https

defaultManagerSettings :: ManagerSettings

newManager :: ManagerSettings -> IO Manager