Cardano.Wallet.Jormungandr.Api.Types

Description

Definitions of types used in the Jörmungandr REST API.

Polymorphic Types

newtype ApiT a

Polymorphic wrapper type to put around primitive types and, 3rd party lib types to avoid defining orphan instances and/or, undesirable instances on primitive types. It helps to keep a nice separation of concerns between the API layer and other modules.

Constructors

ApiT

Fields

getApiT :: a

API types

newtype AccountId

Constructors

AccountId

Fields

getAccountId :: ByteString

data AccountState

Constructors

AccountState

Fields

currentBalance :: !(Quantity "lovelace" Word64)

stakePools :: [(PoolId, Quantity "stake-pool-ratio" Word64)]

totalTransactionCount :: !(Quantity "transaction-count" Word64)

data ApiStakeDistribution

Constructors

ApiStakeDistribution

Fields

dangling :: ApiT (Quantity "lovelace" Word64)

pools :: [(ApiT PoolId, ApiT (Quantity "lovelace" Word64))]

unassigned :: ApiT (Quantity "lovelace" Word64)

newtype BlockId

Constructors

BlockId

Fields

getBlockId :: Hash "BlockHeader"

data StakeApiResponse

Constructors

StakeApiResponse

Fields

epoch :: ApiT EpochNo

stake :: ApiStakeDistribution

Content types

data Hex

data JormungandrBinary

Represents the binary format of Jörmungandr.