Cardano.Wallet.Jormungandr.Binary

Description

The format is for the Shelley era as implemented by the Jörmungandr node. It is described here

The module to some extent defines its own Jörmungandr-specific types, different from Cardano.Wallet.Primitive.Types. Here, transactions are just one of many possible Fragment that can be included in a block.

In some cases it also leads us to throw exceptions when integers would otherwise overflow (look for uses of toEnum).

Documentation

data Block

Constructors

Block

Fields

header :: BlockHeader

fragments :: [Fragment]

data BlockHeader

Constructors

BlockHeader

Fields

version :: Word16

contentSize :: Word32

slot :: SlotId

chainLength :: Word32

contentHash :: Hash "content"

headerHash :: Hash "BlockHeader"

parentHeaderHash :: Hash "BlockHeader"

producedBy :: Maybe PoolId

Will contain the VRFPubKey of the stake pool for non-genesis Genesis/Praos blocks.

data ConfigParam

Constructors

Block0Date StartTime
The official start time of the blockchain, in seconds since the Unix epoch.

Discrimination NetworkDiscriminant
Address discrimination. Testnet / Mainnet.

Consensus ConsensusVersion
Consensus version. BFT / Genesis Praos.

SlotsPerEpoch EpochLength
Number of slots in an epoch.

SlotDuration NominalDiffTime
Slot duration in seconds.

EpochStabilityDepth (Quantity "block" Word32)
The length of the suffix of the chain (in blocks) considered unstable.

ConsensusGenesisPraosParamF Milli
Determines maximum probability of a stakeholder being elected as leader in a slot.

MaxNumberOfTransactionsPerBlock Word32
Maximum number of transactions in a block.

BftSlotsRatio Milli
Fraction of blocks to be created by BFT leaders.

AddBftLeader LeaderId
Add a BFT Leader

RemoveBftLeader LeaderId
Remove a BFT Leader

AllowAccountCreation Bool
Enable/disable account creation.

ConfigLinearFee FeePolicy
Coefficients for fee calculations.

ProposalExpiration (Quantity "epoch" Word32)
Number of epochs until an update proposal expires.

KesUpdateSpeed (Quantity "second/update" Word32)
Maximum number of seconds per update for KES keys known by the system after start time.

ConfigPerCertificate PerCertificateFee
Per certificate fees, override the certificate fee of the linear fee

UnimplementedConfigParam Word16

data ConsensusVersion

Constructors

BFT

GenesisPraos

newtype LeaderId

Constructors

LeaderId ByteString

data Fragment

The block-body consists of fragments. There are several types of fragments.

Constructors

Initial [ConfigParam]
Found in the genesis block.

Transaction Tx
A standard signed transaction

StakeDelegation (StakeDelegationType, ChimericAccount, Tx)
A signed transaction with stake pool delegation

UnimplementedFragment Word8

data FragmentSpec

Constructors

FragmentInitial

FragmentLegacyUTxO

FragmentTransaction

FragmentDelegation

newtype Milli

Used to represent (>= 0) rational numbers as (>= 0) integers, by just multiplying by 1000. For instance: '3.141592' is represented as 'Milli 3142'.

Constructors

Milli Word64

data PerCertificateFee

Constructors

PerCertificateFee

Fields

feePoolRegistration :: Word64

feeStakeDelegation :: Word64

feeOwnerStakeDelegation :: Word64

getBlock :: Get Block

getBlockHeader :: Get BlockHeader

getBlockId :: Get (Hash "BlockHeader")

Extract a Block id from a serialized Block.

getFragment :: Get Fragment

Decode a fragment (header + contents).

Corresponds to FRAGMENT in the specification.

getTransaction :: Hash "Tx" -> Get Tx

Decode the contents of a Transaction-fragment.

overrideFeePolicy :: FeePolicy -> PerCertificateFee -> FeePolicy

Take an existing fee policy and override the certificate using the a given per certificate policy.

This relies on the assumption that the certificate in the FeePolicy is only interpret as a 'stake delegation' certificate.

Constructing Fragment

data MkFragment

Constructors

MkFragmentSimpleTransaction TxWitnessTag

MkFragmentStakeDelegation TxWitnessTag StakeDelegationType ChimericAccount (XPrv, Passphrase "encryption")

data StakeDelegationType

Constructors

DlgNone

DlgFull PoolId

DlgRatio

putFragment :: Hash "Genesis" -> [((TxIn, Coin), (XPrv, Passphrase "encryption"))] -> [TxOut] -> MkFragment -> PutM (Proxy "Fragment")

finalizeFragment :: PutM (Proxy "Fragment") -> SealedTx

fragmentId :: PutM (Proxy "Fragment") -> Hash "Tx"

Transaction witnesses

data TxWitnessTag

Constructors

TxWitnessLegacyUTxO

TxWitnessUTxO

TxWitnessAccount

TxWitnessMultisig

putTxWitnessTag :: TxWitnessTag -> Put

getTxWitnessTag :: Get TxWitnessTag

txWitnessSize :: TxWitnessTag -> Int

Purification of chain block types

convertBlock :: Block -> Block

Convert the Jörmungandr binary format block into a simpler Wallet block.

convertBlockHeader :: BlockHeader -> BlockHeader

Convert the Jörmungandr binary format header into a simpler Wallet header.

Addresses

putAddress :: Address -> Put

getAddress :: Get Address

Helpers

whileM :: Monad m => m Bool -> m a -> m [a]

blake2b256 :: ByteString -> ByteString

Compute a Blake2b_256 hash of a given ByteString

maxNumberOfInputs :: Int

maxNumberOfOutputs :: Int

Re-export

data Get a

runGet :: Get a -> ByteString -> a

The simplest interface to run a Get decoder. If the decoder runs into an error, calls fail, or runs out of input, it will call error.

runGetOrFail :: Get a -> ByteString -> Either (ByteString, ByteOffset, String) (ByteString, ByteOffset, a)

Run a Get monad and return Left on failure and Right on success. In both cases any unconsumed input and the number of bytes consumed is returned. In the case of failure, a human-readable error message is included as well.

Since: 0.6.4.0

type Put = PutM ()

Put merely lifts Builder into a Writer monad, applied to ().

runPut :: Put -> ByteString

Run the PutM monad with a serialiser