Cardano.Wallet.Jormungandr.Compatibility

Description

Contains various implementation decision that are specific to a particular network / protocol. This allows us to easily select a particular backend (Byron, Shelley-Rust, Shelley-Haskell) and isolate the bits that vary between those backends.

Target

data Jormungandr (network :: Network)

A type representing the Jormungandr as a network target. This has an influence on binary serializer & network primitives. See also TxId

data Network

Available network options.

Constructors

Mainnet

Testnet

block0 :: BlockHeader

Genesis block header, i.e. very first block header of the chain

softTxMaxSize :: Quantity "byte" Word16

Jörmugandr's chain parameter doesn't include a transaction max size. The actual hard-limit for the size is constrained by the binary format and numbers used to represent the number of inputs and outputs (Word8), yet there's also a soft-limit of 8kb which results in much smaller transactions in the end.

Node's Configuration

data BaseUrl

Constructors

BaseUrl

Fields

baseUrlScheme :: Scheme

baseUrlHost :: String

baseUrlPort :: Int

baseUrlPath :: String

data Scheme

Constructors

Http

Https

genConfigFile :: FilePath -> Int -> BaseUrl -> Value

Generate a configuration file for Jörmungandr@0.3.1