Cardano.Wallet.Jormungandr.Environment

Description

This module contains static configuration parameters. Rather than providing and carrying around a configuration file through the application, we resolve configuration data at runtime using the available environment.

This gives us a flexible and portable approach to software configuration, and remove some pain from the development perspective. Prior to starting, the wallet is expected to have a few configuration parameter available. One may rely on a `.env` file to bundle configuration settings together for a given target environment.

Networking

data Network

Available network options.

Constructors

Mainnet

Testnet

class KnownNetwork (n :: Network) where

Embed some constants into a network type.

Methods

networkVal :: Network

hrp

Arguments

:: HumanReadablePart
Address discrimination Human-Readable Part (HRP).

"ca" for Mainnet

"ta" for Testnet

single

Arguments

:: Word8
Address discriminant byte for single addresses, this is the first byte of every addresses using the Shelley format carrying only a spending key.

grouped

Arguments

:: Word8
Address discriminant byte for grouped addresses, this is the first byte of every addresses using the Shelley format carrying both a spending and a delegation key.