Cardano.Wallet.Jormungandr.Network

Description

This module allows the wallet to retrieve blocks from a known Jormungandr node. This is done by providing a NetworkLayer with some logic building on top of an underlying JormungandrClient HTTP client.

It also provides facilities for starting the jormungandr node backend process. The Jormungandr NetworkLayer implementation uses the HTTP REST API of this backend.

Starting the network layer

data JormungandrBackend

Whether to start Jormungandr with the given config, or to connect to an already running Jormungandr REST API using the given parameters.

Constructors

UseRunning JormungandrConnParams

Launch JormungandrConfig

data JormungandrConnParams

Parameters for connecting to a Jormungandr REST API.

Constructors

JormungandrConnParams

Fields

_genesisHash :: Hash "Genesis"

_restApi :: BaseUrl

withNetworkLayer

Arguments

:: t ~ Jormungandr

=> Trace IO Text
Logging

-> JormungandrBackend
How Jörmungandr is started.

-> (Either ErrStartup (JormungandrConnParams, NetworkLayer IO t Block) -> IO a)
The action to run. It will be passed the connection parameters used, and a network layer if startup was successful.

-> IO a

Starts the network layer and runs the given action with a NetworkLayer. The caller is responsible for handling errors which may have occurred while starting the Node.

newNetworkLayer :: forall t. t ~ Jormungandr => Trace IO Text -> BaseUrl -> Hash "Genesis" -> ExceptT ErrGetBlockchainParams IO (NetworkLayer IO t Block)

Creates a new NetworkLayer connecting to an underlying Jormungandr backend target.

Launching the node backend

data JormungandrConfig

A subset of the Jormungandr configuration parameters, used for starting the Jormungandr node backend.

Constructors

JormungandrConfig

Fields

_stateDir :: FilePath

_genesisBlock :: Either (Hash "Genesis") FilePath

_restApiPort :: Maybe PortNumber

_outputStream :: StdStream

_extraArgs :: [String]

withJormungandr

Arguments

:: Trace IO Text
Logging

-> JormungandrConfig
Launch configuration

-> (JormungandrConnParams -> IO a)
Action to run while node is running.

-> IO (Either ErrStartup a)

Launches a Jörmungandr node backend with the given configuration

connParamsPort :: JormungandrConnParams -> Int

Extract the port number from the base URL part of the connection params.

Errors

data ErrGetBlock

Error while trying to get one or more blocks

Constructors

ErrGetBlockNetworkUnreachable ErrNetworkUnavailable

ErrGetBlockNotFound (Hash "BlockHeader")

data ErrGetBlockchainParams

Constructors

ErrGetBlockchainParamsNetworkUnreachable ErrNetworkUnavailable

ErrGetBlockchainParamsGenesisNotFound (Hash "Genesis")

ErrGetBlockchainParamsIncompleteParams [ConfigParam]

data ErrGetDescendants

Constructors

ErrGetDescendantsNetworkUnreachable ErrNetworkUnavailable

ErrGetDescendantsParentNotFound (Hash "BlockHeader")

data ErrNetworkTip

Error while trying to get the network tip

Constructors

ErrNetworkTipNetworkUnreachable ErrNetworkUnavailable

ErrNetworkTipNotFound

data ErrNetworkUnavailable

Network is unavailable

Constructors

ErrNetworkUnreachable Text
Cannot connect to network backend.

ErrNetworkInvalid Text
Network backend reports that the requested network is invalid.

data ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxNetworkUnreachable ErrNetworkUnavailable

ErrPostTxBadRequest Text

ErrPostTxProtocolFailure Text

data ErrStartup

Constructors

ErrStartupNodeNotListening

ErrStartupCommandExited ProcessHasExited

ErrStartupGetBlockchainParameters ErrGetBlockchainParams

ErrStartupInvalidGenesisBlock FilePath

ErrStartupInvalidGenesisHash String

data ErrUnexpectedNetworkFailure

Constructors

ErrUnexpectedNetworkFailure Link ServantError

Internal constructors

mkRawNetworkLayer

Arguments

:: (MonadBaseControl IO m, t ~ Jormungandr, block ~ Block)

=> (Block, BlockchainParameters)

-> Word
Batch size when fetching blocks from Jörmungandr

-> MVar BlockHeaders

-> JormungandrClient m

-> NetworkLayer m t block

Wrap a Jormungandr client into a NetworkLayer common interface.

This version provides the full, raw blocks from Cardano.Wallet.Jormungandr.Binary.

data BaseUrl

Constructors

BaseUrl

Fields

baseUrlScheme :: Scheme

baseUrlHost :: String

baseUrlPort :: Int

baseUrlPath :: String

data Scheme

Constructors

Http

Https