Cardano.Wallet.Jormungandr.Rewards

Documentation

data RewardFormula

Constructors

HalvingFormula RewardParams

LinearFormula RewardParams

data RewardParams

Constructors

RewardParams

Fields

rFixed :: Word64

In the linear formula, it represents the starting point of the contribution at #epoch=0, whereas in halving formula is used as starting constant for the calculation.

rRatio :: Ratio Word64

In the halving formula, an effective value between 0.0 to 1.0 indicates a reducing contribution, whereas above 1.0 it indicate an acceleration of contribution.

However in linear formula the meaning is just a scaling factor for the epoch zone (current_epoch - start_epoch / epoch_rate). Further requirement is that this ratio is expressed in fractional form (e.g. 1/2), which allow calculation in integer form.

rEpochStart :: Word32

indicates when this contribution start. note that if the epoch is not the same or after the epoch_start, the overall contribution is zero.

rEpochRate :: Word32

the rate at which the contribution is tweaked related to epoch.

data TaxParameters

Constructors

TaxParameters

Fields

taxFixed :: Word64

A fix value taken from the total

taxRatio :: Ratio Word64

An extra percentage taken from the total

taxLimit :: Maybe Word64

It is possible to add a max bound to the total value taken at each cut.

data RewardLimit

limit the epoch total reward drawing limit to a portion of the total active stake of the system.

for example, if set to 10%, the reward drawn will be bounded by the 10% of the total active stake.

this value is optional, the default is no reward drawing limit

Constructors

RewardLimitNone

RewardLimitByAbsoluteStake (Ratio Word64)

data PoolCapping

settings to incentivize the numbers of stake pool to be registered on the blockchain.

These settings does not prevent more stake pool to be added. For example if there is already 1000 stake pools, someone can still register a new stake pool and affect the rewards of everyone else too.

if the threshold is reached, the pool doesn't really have incentive to create more blocks than 1 / set-value-of-pools % of stake.

this value is optional, the default is no pool participation capping

Constructors

PoolCapping

Fields

minParticipation :: Word32

maxParticipation :: Word32

rewardsAt :: (RewardLimit, Quantity "lovelace" Word64) -> TaxParameters -> EpochNo -> RewardFormula -> Quantity "lovelace" Word64