Cardano.Wallet.Jormungandr

Description

This module provides the main wallet server function for the Jörmungandr backend.

The Cardano.Wallet.Jormungandr.Network module handles synchronization with the jormungandr process.

The block format decoders are in Cardano.Wallet.Jormungandr.Network.

Functionality specific to this backend for creating transactions is in Cardano.Wallet.Jormungandr.Transaction.

Documentation

serveWallet

Arguments

:: forall (n :: NetworkDiscriminant). (t ~ Jormungandr, NetworkDiscriminantVal n, DecodeAddress n, EncodeAddress n, DelegationAddress n ShelleyKey)

=> Tracers IO
Logging config.

-> SyncTolerance
A time tolerance within we consider being synced

-> Maybe FilePath
Database folder filepath

-> HostPreference
Which host to bind.

-> Listen
HTTP API Server port.

-> JormungandrBackend
Whether and how to launch or use the node backend.

-> (SockAddr -> Port "node" -> BlockchainParameters -> IO ())
Callback to run before the main loop

-> IO ExitCode

The cardano-wallet-jormungandr main function. It takes the configuration which was passed from the CLI and environment and starts all components of the wallet.

Utilities

toSPBlock :: Block -> Block

Covert a raw block to one that the Cardano.Pool.Metrics module accepts.

Tracing

data Tracers' f

The types of trace events produced by the Jörmungandr API server.

Constructors

Tracers

Fields

applicationTracer :: f ApplicationLog

apiServerTracer :: f ApiLog

walletEngine :: f WorkerRegistryLog

walletDbTracer :: f DBLog

networkTracer :: f NetworkLayerLog

stakePoolMonitorTracer :: f StakePoolMonitorLog

stakePoolLayerTracer :: f StakePoolLayerLog

stakePoolDBTracer :: f DBLog

daedalusIPCTracer :: f DaedalusIPCLog

type Tracers m = Tracers' (Tracer m)

All of the Jörmungandr Tracers.

type TracerSeverities = Tracers' (Const (Maybe Severity))

The minimum severities for Tracers'. Nothing indicates that tracing is completely disabled.

tracerLabels :: Tracers' (Const Text)

Strings that the user can refer to tracers by.

tracerDescriptions :: [(String, String)]

Names and descriptions of the tracers, for user documentation.

nullTracers :: Monad m => Tracers m

Use a nullTracer for each of the Tracers in Tracers'

setupTracers :: TracerSeverities -> Trace IO Text -> Tracers IO

Set up tracing with textual log messages.

tracerSeverities :: Maybe Severity -> TracerSeverities

Construct a TracerSeverities record with all tracers set to the given severity.