Cardano.Wallet.Jormungandr

Description

This module provides the main wallet server function for the Jörmungandr backend.

The Cardano.Wallet.Jormungandr.Network module handles synchronization with the jormungandr process.

The block format decoders are in Cardano.Wallet.Jormungandr.Network.

Functionality specific to this backend for creating transactions is in Cardano.Wallet.Jormungandr.Transaction.

Documentation

serveWallet

Arguments

:: (n ~ Testnet, t ~ Jormungandr n, s ~ SeqState t, k ~ SeqKey)

=> (Configuration, Switchboard Text, Trace IO Text)
Logging config.

-> Maybe FilePath
Database folder filepath

-> Listen
HTTP API Server port.

-> JormungandrBackend
Whether and how to launch or use the node backend.

-> (Port "wallet" -> Port "node" -> BlockchainParameters -> IO ())
Callback to run before the main loop

-> IO ExitCode

The cardano-wallet-jormungandr main function. It takes the configuration which was passed from the CLI and environment and starts all components of the wallet.