Cardano.Wallet.Jormungandr.Binary

Description

The format is for the Shelley era as implemented by the Jörmungandr node. It is described here

The module to some extent defines its own Jörmungandr-specific types, different from Cardano.Wallet.Primitive.Types. Here, transactions are just one of many possible Messages that can be included in a block.

In some cases it also leads us to throw exceptions when integers would otherwise overflow (look for uses of toEnum).

Documentation

data Block

Constructors

Block BlockHeader [Message]

data BlockHeader

Constructors

BlockHeader

Fields

version :: Word16

contentSize :: Word32

slot :: SlotId

chainLength :: Word32

contentHash :: Hash "content"

parentHeaderHash :: Hash "BlockHeader"

data ConfigParam

Constructors

Block0Date StartTime
The official start time of the blockchain, in seconds since the Unix epoch.

Discrimination Network
Address discrimination. Testnet / Mainnet.

Consensus ConsensusVersion
Consensus version. BFT / Genesis Praos.

SlotsPerEpoch EpochLength
Number of slots in an epoch.

SlotDuration NominalDiffTime
Slot duration in seconds.

EpochStabilityDepth (Quantity "block" Word32)
The length of the suffix of the chain (in blocks) considered unstable.

ConsensusGenesisPraosParamF Milli
Determines maximum probability of a stakeholder being elected as leader in a slot.

MaxNumberOfTransactionsPerBlock Word32
Maximum number of transactions in a block.

BftSlotsRatio Milli
Fraction of blocks to be created by BFT leaders.

AddBftLeader LeaderId
Add a BFT Leader

RemoveBftLeader LeaderId
Remove a BFT Leader

AllowAccountCreation Bool
Enable/disable account creation.

ConfigLinearFee FeePolicy
Coefficients for fee calculations.

ProposalExpiration (Quantity "epoch" Word32)
Number of epochs until an update proposal expires.

KesUpdateSpeed (Quantity "second/update" Word32)
Maximum number of seconds per update for KES keys known by the system after start time.

data ConsensusVersion

Constructors

BFT

GenesisPraos

newtype LeaderId

Constructors

LeaderId ByteString

data Message

The block-body consists of messages. There are several types of messages.

Following, as closely as possible: https://github.com/input-output-hk/rust-cardano/blob/e0616f13bebd6b908320bddb1c1502dea0d3305a/chain-impl-mockchain/src/message/mod.rs#L22-L29

Constructors

Initial [ConfigParam]
Found in the genesis block.

Transaction (Tx, [TxWitness])
A standard signed transaction

UnimplementedMessage Int

data MessageType

Constructors

MsgTypeInitial

MsgTypeTransaction

newtype Milli

Used to represent (>= 0) rational numbers as (>= 0) integers, by just multiplying by 1000. For instance: '3.141592' is represented as 'Milli 3142'.

Constructors

Milli Word64

getBlock :: Get Block

getBlockHeader :: Get BlockHeader

getBlockId :: Get (Hash "BlockHeader")

Extract a Block id from a serialized Block.

getMessage :: Get Message

Decode a message (header + contents).

getTransaction :: Int -> Get (Tx, [TxWitness])

Decode the contents of a Transaction-message.

putSignedTx :: [(TxIn, Coin)] -> [TxOut] -> [TxWitness] -> Put

putTx :: [(TxIn, Coin)] -> [TxOut] -> Put

Coercion with business-domain

coerceBlock :: Block -> Block Tx

Addresses

putAddress :: Address -> Put

getAddress :: Get Address

singleAddressFromKey :: forall n. KnownNetwork n => Proxy n -> XPub -> Address

Legacy Decoders

decodeLegacyAddress :: ByteString -> Maybe Address

Attempt decoding a ByteString into an Address. This merely checks that the underlying bytestring has a "valid" structure / format without doing much more.

Helpers

blake2b256 :: ByteString -> ByteString

Compute a Blake2b_256 hash of a given ByteString

estimateMaxNumberOfInputsParams :: EstimateMaxNumberOfInputsParams t

This provides network encoding specific variables to be used by the estimateMaxNumberOfInputs function.

fragmentId :: [(TxIn, Coin)] -> [TxOut] -> [TxWitness] -> Hash "Tx"

Jörmungandr distinguish 'fragment id' (what we commonly call txId) from 'transaction sign data'. A transaction fragment corresponds to a signed transaction (inputs, outputs and witnesses). So, the witnesses are required to compute a txid.

maxNumberOfInputs :: Int

maxNumberOfOutputs :: Int

signData :: [(TxIn, Coin)] -> [TxOut] -> Hash "SignData"

See fragmentId. This computes the signing data required for producing transaction witnesses.

withHeader :: MessageType -> Put -> Put

Add a corresponding header to a message. Every message is encoded as:

HEADER(MESSAGE) | MESSAGE

where HEADER is:

SIZE (2 bytes) | TYPE (1 byte)

Re-export

data Get a

runGet :: Get a -> ByteString -> a

The simplest interface to run a Get decoder. If the decoder runs into an error, calls fail, or runs out of input, it will call error.

runGetOrFail :: Get a -> ByteString -> Either (ByteString, ByteOffset, String) (ByteString, ByteOffset, a)

Run a Get monad and return Left on failure and Right on success. In both cases any unconsumed input and the number of bytes consumed is returned. In the case of failure, a human-readable error message is included as well.

Since: 0.6.4.0

type Put = PutM ()

Put merely lifts Builder into a Writer monad, applied to ().

runPut :: Put -> ByteString

Run the PutM monad with a serialiser