Cardano.Wallet.Jormungandr.Binary

Description

The format is for the Shelley era as implemented by the Jörmungandr node.

It is described here

Documentation

data Block

Constructors

Block BlockHeader [Message]

data BlockHeader

Constructors

BlockHeader

Fields

version :: Word16

contentSize :: Word32

slot :: SlotId

chainLength :: Word32

contentHash :: Hash "content"

parentHeaderHash :: Hash "BlockHeader"

data Message

The block-body consists of messages. There are several types of messages.

Following, as closely as possible: https://github.com/input-output-hk/rust-cardano/blob/e0616f13bebd6b908320bddb1c1502dea0d3305a/chain-impl-mockchain/src/message/mod.rs#L22-L29

Constructors

Initial [ConfigParam]
Found in the genesis block. | OldUtxoDeclaration UtxoDeclaration

Transaction Tx

UnimplementedMessage Int

getBlockHeader :: Get BlockHeader

getBlock :: Get Block

getTransaction :: Get Tx

Decode the contents of a Transaction-message.

putTokenTransfer :: Tx -> Put

putSignedTransaction :: (Tx, [TxWitness]) -> Put

data ConfigParam

Constructors

Block0Date Word64
The official start time of the blockchain, in seconds since the Unix epoch.

Discrimination Network
Address discrimination. Testnet / Mainnet.

Consensus ConsensusVersion
Consensus version. BFT / Genesis Praos.

SlotsPerEpoch (Quantity "slot/epoch" Word32)
Number of slots in an epoch.

SlotDuration (Quantity "second/slot" Word8)
Slot duration in seconds.

EpochStabilityDepth (Quantity "block" Word32)
The length of the suffix of the chain (in blocks) considered unstable.

ConsensusGenesisPraosParamF Milli
Determines maximum probability of a stakeholder being elected as leader in a slot.

MaxNumberOfTransactionsPerBlock Word32
Maximum number of transactions in a block.

BftSlotsRatio Milli
Fraction of blocks to be created by BFT leaders.

AddBftLeader LeaderId
Add a BFT Leader

RemoveBftLeader LeaderId
Remove a BFT Leader

AllowAccountCreation Bool
Enable/disable account creation.

ConfigLinearFee LinearFee
Coefficients for fee calculations.

ProposalExpiration (Quantity "epoch" Word32)
Number of epochs until an update proposal expires.

KesUpdateSpeed (Quantity "second/update" Word32)
Maximum number of seconds per update for KES keys known by the system after start time.

data ConsensusVersion

Constructors

BFT

GenesisPraos

newtype LeaderId

Constructors

LeaderId ByteString

data LinearFee

Constructors

LinearFee

Fields

constant :: Quantity "lovelace" Word64

perByte :: Quantity "lovelace/byte" Word64

perCert :: Quantity "lovelace/cert" Word64

newtype Milli

Used to represent (>= 0) rational numbers as (>= 0) integers, by just multiplying by 1000. For instance: '3.141592' is represented as 'Milli 3142'.

Constructors

Milli Word64

Addresses

putAddress :: Address -> Put

getAddress :: Get Address

singleAddressFromKey :: forall n. KnownNetwork n => Proxy n -> XPub -> Address

Classes

class FromBinary a where

Methods

get :: Get a

Legacy Decoders

decodeLegacyAddress :: ByteString -> Maybe Address

Attempt decoding a ByteString into an Address. This merely checks that the underlying bytestring has a "valid" structure / format without doing much more.

Re-export

runGet :: Get a -> ByteString -> a

The simplest interface to run a Get decoder. If the decoder runs into an error, calls fail, or runs out of input, it will call error.

data Get a

runPut :: Put -> ByteString

Run the PutM monad with a serialiser

type Put = PutM ()

Put merely lifts Builder into a Writer monad, applied to ().