Cardano.Wallet.Jormungandr.Network

Description

This module allows the wallet to retrieve blocks from a known Jormungandr node. This is done by providing a NetworkLayer with some logic building on top of an underlying JormungandrLayer HTTP client.

Documentation

newNetworkLayer :: forall n. BaseUrl -> IO (NetworkLayer (Jormungandr n) IO)

Creates a new NetworkLayer connecting to an underlying Jormungandr backend target.

mkNetworkLayer :: Monad m => JormungandrLayer n m -> NetworkLayer (Jormungandr n) m

Wrap a Jormungandr client into a NetworkLayer common interface.

data JormungandrLayer n m

Endpoints of the jormungandr REST API.

Constructors

JormungandrLayer

Fields

getTipId :: ExceptT ErrNetworkUnavailable m (Hash "BlockHeader")

getBlock :: Hash "BlockHeader" -> ExceptT ErrGetBlock m Block

getDescendantIds :: Hash "BlockHeader" -> Word -> ExceptT ErrGetDescendants m [Hash "BlockHeader"]

postMessage :: (Tx, [TxWitness]) -> ExceptT ErrPostTx m ()

getInitialBlockchainParameters :: Hash "Genesis" -> ExceptT ErrGetBlockchainParams m (BlockchainParameters (Jormungandr n))

mkJormungandrLayer :: forall n. Manager -> BaseUrl -> JormungandrLayer n IO

Construct a JormungandrLayer-client

>>> mgr <- newManager defaultManagerSettings
>>> j = mkJormungandrLayer mgr (BaseUrl Http "localhost" 8080 "")

>>> (Right tip) <- runExceptT $ getTipId j
>>> tip
BlockId (Hash {getHash = "26c640a3de09b74398c14ca0a137ec78"})

>>> (Right block) <- runExceptT $ getBlock j t
>>> block
>>> Block {header = BlockHeader {slotId = SlotId {epochNumber = 0, slotNumber = 0}, prevBlockHash = Hash {getHash = "\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL"}}, transactions = [Tx {inputs = [], outputs = [TxOut {address = Address {unAddress = "3$\195xi\193\"h\154\&5\145}\245:O\"\148\163\165/h^\ENQ\245\248\229;\135\231\234E/"}, coin = Coin {getCoin = 14}}]}]}

At the time of writing, we only have the genesis-block, but we should be able to get its descendants.

>>> let genesisHash = BlockId (Hash {getHash = "&\198@\163\222\t\183C\152\193L\160\161\&7\236x\245\229\EOT\175\177\167\131\190\b\b/\174\212\177:\179"})
>>> runExceptT $ getDescendantIds j t 4
Right []

Exceptions

data ErrUnexpectedNetworkFailure

Constructors

ErrUnexpectedNetworkFailure Link ServantError

Errors

data ErrGetDescendants

Constructors

ErrGetDescendantsNetworkUnreachable ErrNetworkUnavailable

ErrGetDescendantsParentNotFound (Hash "BlockHeader")

data ErrGetBlockchainParams

Constructors

ErrGetBlockchainParamsNetworkUnreachable ErrNetworkUnavailable

ErrGetBlockchainParamsGenesisNotFound (Hash "Genesis")

ErrGetBlockchainParamsIncompleteParams [ConfigParam]

Re-export

data BaseUrl

Constructors

BaseUrl

Fields

baseUrlScheme :: Scheme

baseUrlHost :: String

baseUrlPort :: Int

baseUrlPath :: String

newManager :: ManagerSettings -> IO Manager

defaultManagerSettings :: ManagerSettings

data Scheme

Constructors

Http

Https

Utils

baseUrlToText :: BaseUrl -> Text

Format an API BaseUrl, for logging, etc.