Cardano.Wallet.Jormungandr.Network

Description

This module allows the wallet to retrieve blocks from a known Jormungandr node. This is done by providing a NetworkLayer with some logic building on top of an underlying JormungandrLayer HTTP client.

It also provides facilities for starting the jormungandr node backend process. The Jormungandr NetworkLayer implementation uses the HTTP REST API of this backend.

Starting the network layer

data JormungandrBackend

Whether to start Jormungandr with the given config, or to connect to an already running Jormungandr REST API using the given parameters.

Constructors

UseRunning JormungandrConnParams

Launch JormungandrConfig

data JormungandrConnParams

Parameters for connecting to a Jormungandr REST API.

Constructors

JormungandrConnParams

Fields

_genesisHash :: Hash "Genesis"

_restApi :: BaseUrl

withNetworkLayer

Arguments

:: Trace IO Text
Logging

-> JormungandrBackend
How Jörmungandr is started.

-> (Either ErrStartup (JormungandrConnParams, NetworkLayer IO Tx (Block Tx)) -> IO a)
The action to run. It will be passed the connection parameters used, and a network layer if startup was successful.

-> IO a

Starts the network layer and runs the given action with a NetworkLayer. The caller is responsible for handling errors which may have occurred while starting the Node.

Launching the node backend

data JormungandrConfig

A subset of the Jormungandr configuration parameters, used for starting the Jormungandr node backend.

Constructors

JormungandrConfig

Fields

_stateDir :: FilePath

_genesisBlock :: FilePath

_secretFile :: FilePath

_restApiPort :: Maybe PortNumber

_minSeverity :: Severity

_outputStream :: StdStream

data JormungandrLayer m

Endpoints of the jormungandr REST API.

Constructors

JormungandrLayer

Fields

getTipId :: ExceptT ErrNetworkUnavailable m (Hash "BlockHeader")

getBlock :: Hash "BlockHeader" -> ExceptT ErrGetBlock m Block

getDescendantIds :: Hash "BlockHeader" -> Word -> ExceptT ErrGetDescendants m [Hash "BlockHeader"]

postMessage :: (Tx, [TxWitness]) -> ExceptT ErrPostTx m ()

getInitialBlockchainParameters :: Hash "Genesis" -> ExceptT ErrGetBlockchainParams m (Block, BlockchainParameters)

getStakeDistribution :: ExceptT ErrNetworkUnavailable m StakeApiResponse

withJormungandr

Arguments

:: Trace IO Text
Logging

-> JormungandrConfig
Launch configuration

-> (JormungandrConnParams -> IO a)
Action to run while node is running.

-> IO (Either ErrStartup a)

Launches a Jörmungandr node backend with the given configuration

connParamsPort :: JormungandrConnParams -> Int

Extract the port number from the base URL part of the connection params.

Errors

data ErrGetBlockchainParams

Constructors

ErrGetBlockchainParamsNetworkUnreachable ErrNetworkUnavailable

ErrGetBlockchainParamsGenesisNotFound (Hash "Genesis")

ErrGetBlockchainParamsIncompleteParams [ConfigParam]

data ErrGetDescendants

Constructors

ErrGetDescendantsNetworkUnreachable ErrNetworkUnavailable

ErrGetDescendantsParentNotFound (Hash "BlockHeader")

data ErrStartup

Constructors

ErrStartupNodeNotListening

ErrStartupGetBlockchainParameters ErrGetBlockchainParams

ErrStartupGenesisBlockFailed FilePath

ErrStartupCommandExited ProcessHasExited

data ErrUnexpectedNetworkFailure

Constructors

ErrUnexpectedNetworkFailure Link ServantError

Re-export

data BaseUrl

Constructors

BaseUrl

Fields

baseUrlScheme :: Scheme

baseUrlHost :: String

baseUrlPort :: Int

baseUrlPath :: String

data Scheme

Constructors

Http

Https

Internal constructors

mkRawNetworkLayer :: MonadBaseControl IO m => (Block, BlockchainParameters) -> MVar UnstableBlocks -> JormungandrLayer m -> NetworkLayer m Tx Block

Wrap a Jormungandr client into a NetworkLayer common interface.

This version provides the full, raw blocks from Cardano.Wallet.Jormungandr.Binary.

mkJormungandrLayer :: Manager -> BaseUrl -> JormungandrLayer IO

Construct a JormungandrLayer-client

>>> mgr <- newManager defaultManagerSettings
>>> j = mkJormungandrLayer mgr (BaseUrl Http "localhost" 8080 "")

>>> (Right tip) <- runExceptT $ getTipId j
>>> tip
BlockId (Hash {getHash = "26c640a3de09b74398c14ca0a137ec78"})

>>> (Right block) <- runExceptT $ getBlock j t
>>> block
>>> Block {header = BlockHeader {slotId = SlotId {epochNumber = 0, slotNumber = 0}, prevBlockHash = Hash {getHash = "\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL"}}, transactions = [Tx {inputs = [], outputs = [TxOut {address = Address {unAddress = "3$\195xi\193\"h\154\&5\145}\245:O\"\148\163\165/h^\ENQ\245\248\229;\135\231\234E/"}, coin = Coin {getCoin = 14}}]}]}

At the time of writing, we only have the genesis-block, but we should be able to get its descendants.

>>> let genesisHash = BlockId (Hash {getHash = "&\198@\163\222\t\183C\152\193L\160\161\&7\236x\245\229\EOT\175\177\167\131\190\b\b/\174\212\177:\179"})
>>> runExceptT $ getDescendantIds j t 4
Right []

Internal state

data UnstableBlocks

A list of block headers and their hashes. The last block in this sequence is the network tip. The first block in this sequence is the block of depth k, which is the last unstable block.

Constructors

UnstableBlocks

Fields

getUnstableBlocks :: !(Seq (Hash "BlockHeader", BlockHeader))

Double-ended queue of block headers, and their IDs.

blockHeight :: !(Quantity "block" Natural)

The block height of the tip of the sequence.

updateUnstableBlocks

Arguments

:: Monad m

=> Natural
Maximum number of unstable blocks (k).

-> m (Hash "BlockHeader")
Fetches tip.

-> (Hash "BlockHeader" -> m (BlockHeader, Quantity "block" Natural))
Fetches block header and its chain height.

-> UnstableBlocks
Current unstable blocks state.

-> m UnstableBlocks

Updates the unstable blocks state using the given "fetch" functions.

This attempts to synchronise the local state with that of the node. The node may be on a different chain to the current unstable blocks, so this function handles switching of chains.

For example, this is what it would do when the local tip is a13, but the node's tip is b15, on a different chain.

                                   local tip ↴
                ┌───┬───  ───┬───┬───┬───┬───┐
 UnstableBlocks │a03│..    ..│a10│a11│a12│a13│
                └───┴───  ───┴───┴───┴───┴───┘
                          ───┬───┬───┬───┬───┬───┬───┐
 Node backend chain       ...│a10│a11│b12│b13│b14│b15│
                          ───┴───┴───┴───┴───┴───┴───┘
                      rollback point ⬏     node tip ⬏

To startBackend with, the node says the tip hash is b15.

Work backwards from tip, fetching blocks and adding them to ac, and removing overlapping blocks from ubs. Overlapping blocks occur when there has been a rollback.

    ubs                     ac
1.  ───┬───┬───┬───┬───┐    ┌───┐
    ...│a10│a11│a12│a13│    │b15│
    ───┴───┴───┴───┴───┘    └───┘
2.  ───┬───┬───┬───┬───┐┌───┬───┐
    ...│a10│a11│a12│a13││b14│b15│
    ───┴───┴───┴───┴───┘└───┴───┘
3.  ───┬───┬───┬───┐┌───┬───┬───┐
    ...│a10│a11│a12││b13│b14│b15│
    ───┴───┴───┴───┘└───┴───┴───┘
4.  ───┬───┬───┐┌───┬───┬───┬───┐
    ...│a10│a11││b12│b13│b14│b15│
    ───┴───┴───┘└───┴───┴───┴───┘

Stop once ubs and ac meet with a block which has the same hash. If they never meet, stop cleanupConfig fetching k blocks.

Finally, to get the new UnstableBlocks, append ac to ubs, and limit the length to k.

The new block height is the height of the first block that was fetched.

If any errors occur while this process is running (for example, fetching a block which has been rolled back and lost from the node's state), it will immediately terminate.

emptyUnstableBlocks :: UnstableBlocks